LPS-56048 Move TrashHandlers into module
