+++ portal-test-internal/src/com/liferay/portal/service/test/PortalRegisterTestUtil.java
@@ -15,7 +15,6 @@
 package com.liferay.portal.service.test;
 
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
-import com.liferay.portal.kernel.trash.TrashHandlerRegistryUtil;
 import com.liferay.portal.kernel.workflow.WorkflowHandlerRegistryUtil;
 import com.liferay.portlet.asset.AssetRendererFactoryRegistryUtil;
 import com.liferay.portlet.asset.model.AssetRendererFactory;
@@ -26,9 +25,6 @@ import com.liferay.portlet.documentlibrary.asset.DLFileEntryAssetRendererFactory
 import com.liferay.portlet.documentlibrary.asset.DLFolderAssetRendererFactory;
 import com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer;
 import com.liferay.portlet.documentlibrary.util.DLFolderIndexer;
-import com.liferay.portlet.messageboards.trash.MBCategoryTrashHandler;
-import com.liferay.portlet.messageboards.trash.MBMessageTrashHandler;
-import com.liferay.portlet.messageboards.trash.MBThreadTrashHandler;
 import com.liferay.portlet.messageboards.util.MBMessageIndexer;
 import com.liferay.portlet.trash.util.TrashIndexer;
 import com.liferay.portlet.usersadmin.util.ContactIndexer;
@@ -83,18 +79,6 @@ public class PortalRegisterTestUtil {
 		_indexersRegistered = true;
 	}
 
-	protected static void registerTrashHandlers() {
-		if (_trashHandlersRegistered) {
-			return;
-		}
-
-		TrashHandlerRegistryUtil.register(new MBCategoryTrashHandler());
-		TrashHandlerRegistryUtil.register(new MBMessageTrashHandler());
-		TrashHandlerRegistryUtil.register(new MBThreadTrashHandler());
-
-		_trashHandlersRegistered = true;
-	}
-
 	protected static void registerWorkflowHandlers() {
 		if (_workflowHandlersRegistered) {
 			return;
@@ -112,7 +96,6 @@ public class PortalRegisterTestUtil {
 
 	private static boolean _assetRendererFactoriesRegistered;
 	private static boolean _indexersRegistered;
-	private static boolean _trashHandlersRegistered;
 	private static boolean _workflowHandlersRegistered;
 
 }
