+++ portal-test-internal/src/com/liferay/portal/service/test/ServiceTestUtil.java
@@ -260,10 +260,6 @@ public class ServiceTestUtil {
 			_log.error(e, e);
 		}
 
-		// Trash
-
-		PortalRegisterTestUtil.registerTrashHandlers();
-
 		// Workflow
 
 		PortalRegisterTestUtil.registerWorkflowHandlers();
